order list feature + code and design fixes

+++ app/src/main/java/com/nalepa/pizzeriaapplication/data/pizza/PizzaSize.kt
@@ -2,7 +2,7 @@ package com.nalepa.pizzeriaapplication.data.pizza
 
 data class PizzaSize(
     var name: String = "",
-    var price: Int = 0,
+    var price: Double = 0.0,
     var diameter: Int = 0
 )
